feat(admin): nueva página /admin/clients con fidelización y tarjetas
Consolida fidelización y gestión de clientes en una nueva sección
dedicada. SettingsPage pierde las tabs Fidelización y Clientes.
BottomNav añade Clientes para owner/admin; barber no lo ve.
